+++ R/plot_inp.R
@@ -1,17 +1,17 @@
 #' @title Plot a swmm model structure
-#' @description This function reads an object of class 'swmm_inp' and extracts
+#' @description This function reads an object of class 'inp' and extracts
 #' subcatchments, links and junctions. All objects are converted to simple feautures
 #' geometries and passed to \code{\link[ggplot2]{geom_sf}}.
-#' @param x An swmm_inp object
+#' @param x An inp object
 #' @param ... currently ignored
 #' @examples  
 #' \dontrun{
-#' swmm_inp <- read_inp("model.inp")
-#' plot(swmm_inp)
+#' inp <- read_inp("model.inp")
+#' plot(inp)
 #' }
 #' @rdname plot
 #' @export
-plot.swmm_inp <- function(x,...) {
+plot.inp <- function(x,...) {
   
   # create list with simple feature objects for 
   # subcatchments, junctions and nodes
@@ -19,7 +19,6 @@ plot.swmm_inp <- function(x,...) {
   ggplot2::ggplot() + 
     ggplot2::geom_sf(data = sff$subcatchments) + 
     ggplot2::geom_sf(data = sff$links) + 
-    ggplot2::geom_sf(data = sff$junctions) + 
-    ggplot2::theme_bw()  
+    ggplot2::geom_sf(data = sff$junctions)
   
 }