+++ R/read_inp.R
@@ -110,8 +110,106 @@ read_inp <- function(inp, rm.comment = TRUE) {
   # rename chunks
   names(list_inp) <- tolower(section_title)
   
-  # assign class "swmm_inp"
-  class(list_inp) <- "swmm_inp"
+  # assign class "inp"
+  class(list_inp) <- "inp"
   
+  # change to tibble
+  #list_inp <- lapply(list_inp, tibble::as_tibble)
+    
   return(list_inp)
+}
+
+#' Read SWMM's .inp file
+#'
+#' Reads a SWMM .inp file and creates a list with corresponding SWMM sections.
+#' 
+#' @param inp Name (incl. path) to an input file.
+#' @param rm.comment Should lines with comments starting with a ";" be discarded?
+#' @return An object of class `inp`
+#' @examples  
+#' \dontrun{
+#' inp_sections <- read_inp("model.inp")
+#' } 
+#' @rdname read_inp2
+#' @export 
+read_inp2 <- function(x, rm.comment = TRUE) {
+  
+  #x <- "../../GitLab/models/residential/qm_rc_tmpl.inp"
+  
+  # read lines
+  inp_lines <- readLines(x)
+  
+  # find section lines
+  section_lines <- grep("\\[", inp_lines, value = F)
+  
+  # get section names
+  section_types <- gsub(pattern = "[[:punct:]]", 
+                        replacement = "",  
+                        x = inp_lines[section_lines])
+  
+  # get last line per section
+  section_length <- c(section_lines[-1]-2,length(inp_lines)) 
+  
+  # create list with sections  
+  list_of_sections <- purrr::map2(section_lines, section_length, 
+                                  ~ inp_lines[(.x + 1):.y]) %>% 
+    purrr::set_names(base::tolower(section_types))
+  
+  # parse sections individually
+  res <- purrr::imap(list_of_sections, ~ parse_section(.x, .y, rm.comment = rm.comment)) %>% 
+    # discard nulls to keep list as simple as possible
+    purrr::discard(is.null)
+  
+  # assign class attribute
+  class(res) <- "inp"
+  
+  return(res)
+  
+}
+
+#' import helper
+#' @keywords internal
+parse_section <- function(x, section_type, rm.comment) {
+  
+  # remove header lines 
+  x <- x[!startsWith(x, ";;")]
+  
+  # remove comments
+  if (rm.comment) x <- x[!startsWith(x, ";")]
+  
+  # convert character vector to tibble
+  x <- tibble::as_tibble(x)
+  
+  # switch between section_types
+  x <- switch (section_type,
+               "options" = import_options(x),
+               "title" = import_title(x),
+               "raingages" = import_raingages(x),
+               "subcatchments" = import_subcatchments(x),
+               "subareas" = import_subareas(x),
+               "infiltration" = import_infiltration(x),
+               "junctions" = import_junctions(x),
+               "conduits" = import_conduits(x),
+               "pollutants" = import_pollutants(x),
+               "landuses" = import_landuses(x),
+               "buildup" = import_buildup(x),
+               "washoff" = import_washoff(x),
+               "coordinates" = import_coordinates(x),
+               "vertices" = import_vertices(x),
+               "polygons" = import_polygons(x),
+               "symbols" = import_symbols(x),
+               "labels" = import_labels(x)
+  )
+  
+  # if a section is not parsed, we yield NULL
+  if (is.null(x)) return (NULL)
+  
+  # remove dummy columns which names starts with *tab 
+  x <- x[, !grepl("^tab", colnames(x))]
+  
+  # remove rows with NA's only
+  x <- x[rowSums(is.na(x)) != ncol(x), ]
+  
+  # always return a tibble
+  return(x)
 }